+++ packages/backend/src/storage/MessagesAccessController.ts
@@ -1,8 +1,9 @@
 import AccessController from 'orbit-db-access-controllers'
 import { getCrypto } from 'pkijs'
 import { stringToArrayBuffer } from 'pvutils'
-import { ChannelMessage } from '@quiet/state-manager'
 import { keyObjectFromString, verifySignature } from '@quiet/identity'
+import { ChannelMessage, NoCryptoEngineError } from '@quiet/types'
+import OrbitDB from 'orbit-db'
 
 const type = 'messagesaccess'
 
@@ -16,12 +17,14 @@ export class MessagesAccessController extends AccessController {
     return type
   }
 
-  async canAppend(entry) {
+  async canAppend(entry: LogEntry<ChannelMessage>) {
+    if (!this.crypto) throw new NoCryptoEngineError()
+
     const message: ChannelMessage = entry.payload.value
 
     const signature = stringToArrayBuffer(message.signature)
 
-    let cryptoKey = this.keyMapping[message.pubKey]
+    let cryptoKey = this.keyMapping.get(message.pubKey)
     if (!cryptoKey) {
       cryptoKey = await keyObjectFromString(message.pubKey, this.crypto)
       this.keyMapping.set(message.pubKey, cryptoKey)
@@ -39,7 +42,7 @@ export class MessagesAccessController extends AccessController {
     return ''
   }
 
-  static create(_orbitdb, _type = type, _options) {
+  static create(_orbitdb: OrbitDB, _type = type, _options: any) {
     return new MessagesAccessController()
   }
 }