+++ packages/backend/src/storage/ipfsFileManager.ts
@@ -4,7 +4,7 @@ import path from 'path'
 
 import PQueue, { AbortError } from 'p-queue'
 
-import { decode } from '@ipld/dag-pb'
+import { decode, PBNode } from '@ipld/dag-pb'
 
 import * as base58 from 'multiformats/bases/base58'
 
@@ -17,15 +17,10 @@ import sizeOf from 'image-size'
 
 import { CID } from 'multiformats/cid'
 
-import {
-    FileMetadata,
-    DownloadStatus,
-    DownloadProgress,
-    DownloadState,
-    imagesExtensions
-} from '@quiet/state-manager'
 import { sleep } from '../sleep'
-
+import logger from '../logger'
+import { DownloadProgress, DownloadState, DownloadStatus, FileMetadata, imagesExtensions } from '@quiet/types'
+const log = logger('ipfsFiles')
 const sizeOfPromisified = promisify(sizeOf)
 
 const { createPaths, compare } = await import('../common/utils')
@@ -170,18 +165,18 @@ export class IpfsFilesManager extends EventEmitter {
     }
 
     public async uploadFile(metadata: FileMetadata) {
-        let width: number = null
-        let height: number = null
+        let width: number | undefined
+        let height: number | undefined
         if (imagesExtensions.includes(metadata.ext)) {
-            let imageSize = null
+            let imageSize: {width: number | undefined; height: number | undefined} | undefined // ISizeCalculationResult
             try {
                 imageSize = await sizeOfPromisified(metadata.path)
             } catch (e) {
                 console.error(`Couldn't get image dimensions (${metadata.path}). Error: ${e.message}`)
                 throw new Error(`Couldn't get image dimensions (${metadata.path}). Error: ${e.message}`)
             }
-            width = imageSize.width
-            height = imageSize.height
+            width = imageSize?.width
+            height = imageSize?.height
         }
 
         const stream = fs.createReadStream(metadata.path, { highWaterMark: 64 * 1024 * 10 })
@@ -258,8 +253,8 @@ export class IpfsFilesManager extends EventEmitter {
         }
     }
 
-    private getLocalBlocks = async () => {
-        const blocks = []
+    private getLocalBlocks = async (): Promise<string[]> => {
+        const blocks: string[] = []
 
         const refs = this.ipfs.refs.local()
 
@@ -275,7 +270,7 @@ export class IpfsFilesManager extends EventEmitter {
         const block = CID.parse(fileMetadata.cid)
 
         const localBlocks = await this.getLocalBlocks()
-        const processedBlocks = []
+        const processedBlocks: PBNode[] = [] // TODO: Should it be CID or PBNode?
 
         const controller = new AbortController()
 
@@ -307,8 +302,13 @@ export class IpfsFilesManager extends EventEmitter {
             }
         }
 
+        interface BlockStat {
+            fetchTime: number
+            byteLength: number
+        }
+
         // Transfer speed
-        const blocksStats = []
+        const blocksStats: BlockStat[] = []
 
         const updateTransferSpeed = setInterval(async () => {
             const bytesDownloaded = blocksStats.reduce((previousValue, currentValue) => {
@@ -325,6 +325,10 @@ export class IpfsFilesManager extends EventEmitter {
             }, 0)
             const transferSpeed = bytesDownloaded === 0 ? 0 : bytesDownloaded / TRANSFER_SPEED_SPAN
             const fileState = this.files.get(fileMetadata.cid)
+            if (!fileState) {
+                log.error(`No saved data for file cid ${fileMetadata.cid}`)
+                return
+            }
             this.files.set(fileMetadata.cid, {
                 ...fileState, transferSpeed: transferSpeed, downloadedBytes: totalBytesDownloaded
             })
@@ -354,6 +358,7 @@ export class IpfsFilesManager extends EventEmitter {
                 if (signal.aborted) onAbort()
                 signal.addEventListener('abort', onAbort, { once: true })
 
+                // @ts-ignore FIXME
                 if (processedBlocks.includes(block)) {
                     remainingBlocks.delete(block)
                     resolve(block)
@@ -372,7 +377,7 @@ export class IpfsFilesManager extends EventEmitter {
                     return
                 }
 
-                const decodedBlock = decode(fetchedBlock)
+                const decodedBlock: PBNode = decode(fetchedBlock)
 
                 const fileState = this.files.get(fileMetadata.cid)
 
@@ -386,7 +391,7 @@ export class IpfsFilesManager extends EventEmitter {
                 if (!hasBlockBeenDownloaded) {
                     blocksStats.push({
                         fetchTime: Math.floor(Date.now() / 1000),
-                        byteLength: decodedBlock.Data.byteLength
+                        byteLength: decodedBlock.Data?.byteLength || 0
                     })
                 }
 
@@ -408,9 +413,13 @@ export class IpfsFilesManager extends EventEmitter {
 
         clearInterval(updateTransferSpeed)
 
-        if (this.cancelledDownloads.has(fileMetadata.cid)) {
-            const fileState = this.files.get(fileMetadata.cid)
+        const fileState = this.files.get(fileMetadata.cid)
+        if (!fileState) {
+            log.error(`No saved data for file cid ${fileMetadata.cid}`)
+            return
+        }
 
+        if (this.cancelledDownloads.has(fileMetadata.cid)) {
             this.files.set(fileMetadata.cid, {
                 ...fileState, downloadedBytes: 0, transferSpeed: 0
             })
@@ -419,7 +428,6 @@ export class IpfsFilesManager extends EventEmitter {
             await this.updateStatus(fileMetadata.cid, DownloadState.Canceled)
             this.files.delete(fileMetadata.cid)
         } else {
-            const fileState = this.files.get(fileMetadata.cid)
             this.files.set(fileMetadata.cid, {
                 ...fileState, transferSpeed: 0
             })
@@ -468,7 +476,11 @@ export class IpfsFilesManager extends EventEmitter {
 
     private updateStatus = async (cid: string, downloadState = DownloadState.Downloading) => {
         const metadata = this.files.get(cid)
-        const progress: DownloadProgress = downloadState !== DownloadState.Malicious ? {
+        if (!metadata) {
+          // TODO: emit error?
+          return
+        }
+        const progress: DownloadProgress | undefined = downloadState !== DownloadState.Malicious ? {
             size: metadata.size,
             downloaded: metadata.downloadedBytes,
             transferSpeed: metadata.transferSpeed