+++ packages/desktop/src/renderer/components/ChannelCreationModal/ChannelCreationModal.tsx
@@ -1,21 +1,10 @@
-import React, { useEffect } from 'react'
-import { useSelector } from 'react-redux'
+import React from 'react'
 import { useModal } from '../../containers/hooks'
 import { ModalName } from '../../sagas/modals/modals.types'
-import { publicChannels } from '@quiet/state-manager'
 import ChannelCreationModalComponent from './ChannelCreationModal.component'
 
 const ChannelCreationModal = () => {
   const channelCreationModal = useModal(ModalName.channelCreationModal)
-  const pendingGeneralChannelRecreation = useSelector(publicChannels.selectors.pendingGeneralChannelRecreation)
-
-  useEffect(() => {
-    if (pendingGeneralChannelRecreation) {
-      channelCreationModal.handleOpen()
-    } else {
-      channelCreationModal.handleClose()
-    }
-  }, [pendingGeneralChannelRecreation])
 
   return <ChannelCreationModalComponent {...channelCreationModal} />
 }