+++ packages/desktop/src/renderer/components/Sidebar/ChannelsPanel/ChannelsPanel.tsx
@@ -43,6 +43,7 @@ const ChannelsPanel: React.FC<ChannelsPanelProps> = ({
                 selected={selected}
                 setCurrentChannel={setCurrentChannel}
                 key={channel.address}
+                disabled={channel.disabled}
               />
             )
           })}