+++ packages/mobile/src/components/ChannelList/ChannelList.types.ts
@@ -6,6 +6,4 @@ export interface ChannelListProps {
   community: Community
   tiles: ChannelTileProps[]
   communityContextMenu: ReturnType<typeof useContextMenu>
-  deleteChannel: (channel: string) => void
-  enableDeletion?: boolean
 }