+++ packages/mobile/src/screens/Channel/Channel.screen.tsx
@@ -3,15 +3,18 @@ import { BackHandler, Linking } from 'react-native'
 import { useDispatch, useSelector } from 'react-redux'
 import { Chat } from '../../components/Chat/Chat.component'
 import {
+  communities,
+  identity,
+  publicChannels,
+  messages,
+  files,
   CancelDownload,
   FileMetadata,
-  files,
-  identity,
-  messages,
-  publicChannels
 } from '@quiet/state-manager'
 import { navigationActions } from '../../store/navigation/navigation.slice'
 import { ScreenNames } from '../../const/ScreenNames.enum'
+import { useContextMenu } from '../../hooks/useContextMenu'
+import { MenuName } from '../../const/MenuNames.enum'
 
 export const ChannelScreen: FC = () => {
   const dispatch = useDispatch()
@@ -40,6 +43,8 @@ export const ChannelScreen: FC = () => {
   const currentIdentity = useSelector(identity.selectors.currentIdentity)
   const currentChannel = useSelector(publicChannels.selectors.currentChannel)
 
+  const community = useSelector(communities.selectors.currentCommunity)
+
   const channelMessagesCount = useSelector(publicChannels.selectors.currentChannelMessagesCount)
 
   const channelMessages = useSelector(publicChannels.selectors.currentChannelMessagesMergedBySender)
@@ -47,6 +52,11 @@ export const ChannelScreen: FC = () => {
   const pendingMessages = useSelector(messages.selectors.messagesSendingStatus)
 
   const downloadStatusesMapping = useSelector(files.selectors.downloadStatuses)
+
+  let contextMenu = useContextMenu(MenuName.Channel)
+  if (!community?.CA) {
+    contextMenu = null
+  }
 
   const downloadFile = useCallback(
     (media: FileMetadata) => {
@@ -89,6 +99,7 @@ export const ChannelScreen: FC = () => {
 
   return (
     <Chat
+      contextMenu={contextMenu}
       sendMessageAction={sendMessageAction}
       loadMessagesAction={loadMessages}
       handleBackButton={handleBackButton}