+++ packages/state-manager/src/sagas/communities/communities.types.ts
@@ -1,5 +1,3 @@
-import { createRootCA } from '@quiet/identity'
-import { AsyncReturnType } from '../../utils/types/AsyncReturnType.interface'
 import { HiddenService, PeerId, Identity } from '../identity/identity.types'
 import { Community } from './communities.slice'
 
@@ -41,13 +39,6 @@ export interface InitCommunityPayload {
 export interface UpdateCommunityPayload {
   id: string
   rootCa: string
-}
-
-export interface AddNewCommunityPayload {
-  id: string
-  name: string
-  CA: AsyncReturnType<typeof createRootCA> | {}
-  registrarUrl: string
 }
 
 export interface LaunchRegistrarPayload {