+++ packages/mobile/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1-alpha.6](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.2.1-alpha.5...@quiet/mobile@1.2.1-alpha.6) (2023-05-10)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+## [1.2.1-alpha.5](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.2.1-alpha.4...@quiet/mobile@1.2.1-alpha.5) (2023-05-10)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+## [1.2.1-alpha.4](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.2.1-alpha.3...@quiet/mobile@1.2.1-alpha.4) (2023-05-10)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+## [1.2.1-alpha.3](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.2.1-alpha.2...@quiet/mobile@1.2.1-alpha.3) (2023-05-10)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+## [1.2.1-alpha.2](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.2.1-alpha.1...@quiet/mobile@1.2.1-alpha.2) (2023-04-28)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
 ## [1.2.1-alpha.1](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.2.1-alpha.0...@quiet/mobile@1.2.1-alpha.1) (2023-04-27)
 
 **Note:** Version bump only for package @quiet/mobile